refactor(search-view): extract useSlashCommands hook
